Fix: rename file
PageControllerからPageControlにファイル名を変更
